refactor: move error/warning routines to `_errors.py`
